% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalplots.R
\name{finalplots}
\alias{finalplots}
\title{Finalplots}
\usage{
finalplots(dfname, final_loc_colname, species1, species2, species3)
}
\arguments{
\item{dfname}{name of data frame}

\item{final_loc_colname}{name of column in data frame}

\item{species1}{denotation of plant species 1 being tested}

\item{species2}{denotation of plant species 2 being tested}

\item{species3}{denotation of plant species 3 being tested}
}
\description{
Takes input as dataframe containing final locations as well as denotations for species identities, returns final location statistics and bar plot.
}
\note{
Please format final_loc_colname as dataframe$colname, and be sure to include quotes around each species denotation.
}
\examples{
finalplots(sampledata, sampledata$Final.Location, "A", "P", "V")
}
\keyword{choice}
\keyword{final}
\keyword{food}
\keyword{location}
