% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movementstats.R
\name{movementstats}
\alias{movementstats}
\title{Movementstats}
\usage{
movementstats(placement_column, check1, check2, check3, check4, final_loc)
}
\arguments{
\item{placement_column}{column of data frame containing locations placement locations}

\item{check1}{column of data frame containing locations at check 1}

\item{check2}{column of data frame containing locations at check 2}

\item{check3}{column of data frame containing locations at check 3}

\item{check4}{column of data frame containing locations at check 4}

\item{final_loc}{column of data frame containing locations after full trial time elapsed}
}
\description{
Takes input as dataframe containing inital placement, check 1, check 2, check 3, check 4, and final locations for a number of food choice trials and returns statistics on how frequently movement occurred between checks.
}
\examples{
movementstats(sampledata$Placement, sampledata$Check.1, sampledata$Check.2, sampledata$Check.3, sampledata$Check.4, sampledata$Final.Location)
}
\keyword{choice}
\keyword{food}
\keyword{movement}
