% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeconv.R
\name{timeconv}
\alias{timeconv}
\title{Timeconv}
\usage{
timeconv(time_column)
}
\arguments{
\item{time_column}{Data frame column containing trial times in hh:mm format}
}
\description{
Takes input as column of dataframe showing trial times in hh:mm format and converts to decimal number of hours for easier further quantatative analyses.
}
\examples{
timeconv(sampledata$Time)
}
\keyword{convert}
\keyword{decimal}
\keyword{time}
\keyword{trial}
